test: split in unit and integration
