+++ tests/integration/telemetry/integration_test/test_integration_azure_monitor.py
@@ -1,16 +1,3 @@
-# Copyright 2020 Energinet DataHub A/S
-#
-# Licensed under the Apache License, Version 2.0 (the "License2");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 import sys
 import time
 import uuid
@@ -22,7 +9,9 @@ from azure.monitor.query import LogsQueryClient, LogsQueryPartialResult, LogsQue
 
 import opengeh_utilities.telemetry.logging_configuration as config
 from opengeh_utilities.telemetry.logger import Logger
-from tests.telemetry.integration_test_configuration import IntegrationTestConfiguration
+from tests.integration.telemetry.integration_test_configuration import (
+    IntegrationTestConfiguration,
+)
 
 INTEGRATION_TEST_LOGGER_NAME = "test-logger"
 INTEGRATION_TEST_CLOUD_ROLE_NAME = "test-cloud-role-name"
