--- tests/testing/covernator/constants.py
@@ -1,3 +1,0 @@
-from tests.constants import TEST_DIR
-
-COVERNATOR_TEST_DATA = TEST_DIR / "testing" / "covernator" / "data"
