--- tests/testing/etl/constants.py
@@ -1,3 +1,0 @@
-from tests.constants import TEST_DIR
-
-ETL_TEST_DATA = TEST_DIR / "testing" / "etl" / "data"
