+++ tests/unit/telemetry/telemetry_logging/test_span_recording.py
