+++ tests/unit/testing/covernator/constants.py
@@ -1,0 +1,3 @@
+from tests.constants import TEST_DIR
+
+COVERNATOR_TEST_DATA = TEST_DIR / "unit" / "testing" / "covernator" / "data"
