+++ tests/unit/testing/etl/constants.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+ETL_TEST_DATA = Path(__file__).parent / "data"
