+++ tests/unit/testing/etl/test_test_cases.py
@@ -1,7 +1,7 @@
 from pyspark.sql import types as T
 
 from opengeh_utilities.testing.etl.TestCases import TestCase, TestCases
-from tests.testing.etl.constants import ETL_TEST_DATA
+from tests.unit.testing.etl.constants import ETL_TEST_DATA
 
 
 def test_test_cases(spark):
